Removed internal references from public headers
