+++ sdk/core/azure-core-tracing-opentelemetry/CMakeLists.txt
@@ -41,6 +41,7 @@ if (BUILD_AZURE_CORE_TRACING_OPENTELEMETRY)
     set(
       AZURE_CORE_OPENTELEMETRY_HEADER
         inc/azure/core/tracing/opentelemetry/opentelemetry.hpp
+        src/opentelemetry_impl.hpp
     )
 
     set(
@@ -79,13 +80,13 @@ if(BUILD_AZURE_CORE_TRACING_OPENTELEMETRY)
   az_vcpkg_export(
     azure-core-tracing-opentelemetry
     CORE_TRACING_OPENTELEMETRY
-    "azure/core/tracing/opentelemetry/dll_import_export.hpp"
+    azure/core/tracing/opentelemetry/dll_import_export.hpp
   )
 
   az_rtti_setup(
     azure-core-tracing-opentelemetry
     CORE_TRACING_OPENTELEMETRY
-    "azure/core/tracing/opentelemetry/rtti.hpp"
+    azure/core/tracing/opentelemetry/rtti.hpp
   )
 endif()
