+++ sdk/core/azure-core-tracing-opentelemetry/test/ut/service_support_test.cpp
@@ -300,18 +300,18 @@ TEST_F(OpenTelemetryServiceTests, CreateWithExplicitProvider)
   //
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
     Azure::Core::Context rootContext;
-    rootContext.SetTracerProvider(provider);
+    rootContext.SetTracerProvider(std::static_pointer_cast<Azure::Core::Tracing::TracerProvider>(provider));
     EXPECT_EQ(provider, rootContext.GetTracerProvider());
   }
 
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
     // Create a serviceTrace and span using a provider specified in the ClientOptions.
     {
@@ -349,10 +349,10 @@ TEST_F(OpenTelemetryServiceTests, CreateWithImplicitProvider)
 {
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
-    Azure::Core::Context::ApplicationContext.SetTracerProvider(provider);
+    Azure::Core::Context::ApplicationContext.SetTracerProvider(std::static_pointer_cast<Azure::Core::Tracing::TracerProvider>(provider));
 
     {
       Azure::Core::_internal::ClientOptions clientOptions;
@@ -392,8 +392,8 @@ TEST_F(OpenTelemetryServiceTests, CreateSpanWithOptions)
 {
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
     Azure::Core::Context::ApplicationContext.SetTracerProvider(provider);
 
@@ -441,8 +441,8 @@ TEST_F(OpenTelemetryServiceTests, NestSpans)
 {
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
     Azure::Core::Context::ApplicationContext.SetTracerProvider(provider);
 
@@ -662,8 +662,8 @@ TEST_F(OpenTelemetryServiceTests, ServiceApiImplementation)
 {
   {
     auto tracerProvider(CreateOpenTelemetryProvider());
-    auto provider(std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-        tracerProvider));
+    auto provider(
+        Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(tracerProvider));
 
     {
       // Call a simple API and verify telemetry is generated.
