+++ sdk/core/azure-core/CMakeLists.txt
@@ -89,6 +89,7 @@ set(
     inc/azure/core/internal/json/json_serializable.hpp
     inc/azure/core/internal/strings.hpp
     inc/azure/core/internal/tracing/service_tracing.hpp
+    inc/azure/core/internal/tracing/tracing_impl.hpp
     inc/azure/core/io/body_stream.hpp
     inc/azure/core/match_conditions.hpp
     inc/azure/core/modified_conditions.hpp
