+++ sdk/core/azure-core/src/http/user_agent.cpp
@@ -13,7 +13,6 @@
 #include "azure/core/http/policies/policy.hpp"
 #include "azure/core/internal/tracing/service_tracing.hpp"
 #include "azure/core/platform.hpp"
-#include "azure/core/tracing/tracing.hpp"
 #include <cctype>
 #include <sstream>
