+++ sdk/core/azure-core/test/ut/service_tracing_test.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/tracing/tracing.hpp"
+#include "azure/core/internal/tracing/tracing_impl.hpp"
 #include <azure/core/internal/tracing/service_tracing.hpp>
 #include <gtest/gtest.h>
